@@ -39,11 +39,11 @@ public class LocaleHelper {
     public static Context setLocale(Context context, String language) {
         persist(context, language);
 
-        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.N) {
+//        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.N) {
             return updateResources(context, language);
-        }
-
-        return updateResourcesLegacy(context, language);
+//        } else {
+//            return updateResourcesLegacy(context, language);
+//        }
     }
 
     private static String getPersistedData(Context context, String defaultLanguage) {
@@ -71,21 +71,21 @@ public class LocaleHelper {
         return context.createConfigurationContext(configuration);
     }
 
-    @SuppressWarnings("deprecation")
-    private static Context updateResourcesLegacy(Context context, String language) {
-        Locale locale = new Locale(language);
-        Locale.setDefault(locale);
-
-        Resources resources = context.getResources();
-
-        Configuration configuration = resources.getConfiguration();
-        configuration.locale = locale;
-        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.JELLY_BEAN_MR1) {
-            configuration.setLayoutDirection(locale);
-        }
-
-        resources.updateConfiguration(configuration, resources.getDisplayMetrics());
-
-        return context;
-    }
+//    @TargetApi(Build.VERSION_CODES.M)
+//    private static Context updateResourcesLegacy(Context context, String language) {
+//        Locale locale = new Locale(language);
+//        Locale.setDefault(locale);
+//
+//        Resources resources = context.getResources();
+//
+//        Configuration configuration = resources.getConfiguration();
+//        configuration.locale = locale;
+//        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.JELLY_BEAN_MR1) {
+//            configuration.setLayoutDirection(locale);
+//        }
+//
+//        resources.updateConfiguration(configuration, resources.getDisplayMetrics());
+//
+//        return context;
+//    }
 }
