@@ -2,20 +2,18 @@ package com.thing.quoter.adapter
 
 import android.content.Context
 import android.view.LayoutInflater
-import android.view.View
 import android.view.ViewGroup
 import android.widget.ImageView
 import android.widget.TextView
 import androidx.recyclerview.widget.RecyclerView
-import com.thing.quoter.R
-import com.thing.quoter.repository.model.QuoteProvider
+import com.thing.quoter.databinding.ProviderItemBinding
 import com.thing.quoter.helper.FirestoreList
-import kotlinx.android.synthetic.main.provider_item.view.*
+import com.thing.quoter.repository.model.QuoteProvider
 
-class ProviderViewAdapter(val context: Context, val providers: FirestoreList<QuoteProvider>) : RecyclerView.Adapter<ProviderViewAdapter.ProviderViewHolder>() {
+class ProviderViewAdapter(val context: Context, private val providers: FirestoreList<QuoteProvider>) : RecyclerView.Adapter<ProviderViewAdapter.ProviderViewHolder>() {
     override fun onCreateViewHolder(p0: ViewGroup, p1: Int): ProviderViewHolder {
-        val view = LayoutInflater.from(context).inflate(R.layout.provider_item, p0, false)
-        return ProviderViewHolder(view)
+        val binding = ProviderItemBinding.inflate(LayoutInflater.from(context), p0, false)
+        return ProviderViewHolder(binding)
     }
 
     override fun getItemCount(): Int {
@@ -28,9 +26,9 @@ class ProviderViewAdapter(val context: Context, val providers: FirestoreList<Quo
         p0.bind(quoteProvider)
     }
 
-    class ProviderViewHolder(itemView: View) : RecyclerView.ViewHolder(itemView) {
-        var providerImage: ImageView = itemView.providerImage
-        val providerName: TextView = itemView.providerName
+    class ProviderViewHolder(val itemBinding: ProviderItemBinding) : RecyclerView.ViewHolder(itemBinding.root) {
+        var providerImage: ImageView = itemBinding.providerImage
+        val providerName: TextView = itemBinding.providerName
         fun bind(quoteProvider: QuoteProvider) {
             providerName.text = quoteProvider.providerName
 //            Picasso.get()
